+++ src/Engine.interface.ts
@@ -15,7 +15,6 @@ export interface IEngineGetValid {
 }
 
 export interface IEngineUpdate {
-  id: string;
   workingSetSizeMode?: TEngineWorkingSetSizeMode;
   workingSetSizeLoPct?: TRangeOf100;
   workingSetSizeHiPct?: TRangeOf100;
@@ -24,19 +23,29 @@ export interface IEngineUpdate {
   allowDataLineage?: boolean;
   standardReload?: boolean;
   documentDirectory?: string;
-  documentTimeout?: number;
-  autosaveInterval?: number;
-  genericUndoBufferMaxSize?: number;
-  auditActivityLogVerbosity?: TRangeOf5;
   auditSecurityLogVerbosity?: TRangeOf5;
   systemLogVerbosity?: TRangeOf5;
   externalServicesLogVerbosity?: TRangeOf5;
   qixPerformanceLogVerbosity?: TRangeOf5;
-  serviceLogVerbosity?: TRangeOf5;
   httpTrafficLogVerbosity?: TRangeOf5;
   auditLogVerbosity?: TRangeOf5;
   trafficLogVerbosity?: TRangeOf5;
   sessionLogVerbosity?: TRangeOf5;
   performanceLogVerbosity?: TRangeOf5;
   sseLogVerbosity?: TRangeOf5;
+  listenerPorts?: number[];
+  globalLogMinuteInterval?: number;
+  autosaveInterval?: number;
+  documentTimeout?: number;
+  genericUndoBufferMaxSize?: number;
+  auditActivityLogVerbosity?: TRangeOf5;
+  serviceLogVerbosity?: TRangeOf5;
+  qrsHttpNotificationPort?: number;
+  hyperCubeMemoryLimit?: number;
+  reloadMemoryLimit?: number;
+  exportMemoryLimit?: number;
+  objectTimeLimitSec?: number;
+  exportTimeLimitSec?: number;
+  reloadTimeLimitSec?: number;
+  createSearchIndexOnReloadEnabled?: boolean;
 }